backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
alembic==1.14.0
psycopg2-binary==2.9.10

# Settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Optimization
numpy==2.1.3
scipy==1.14.1

# Reports
openpyxl==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
